% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burned_annot.R
\name{burned_annot}
\alias{burned_annot}
\title{Extract Burned-in Annotation field}
\usage{
burned_annot(file)
}
\arguments{
\item{file}{DICOM file name}
}
\value{
Character vector from header
}
\description{
Reads the DICOM header and extracts the burned in tag:
\code{(0028,0301)}.
}

