% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_tags.R
\docType{data}
\name{date_tags}
\alias{date_tags}
\title{Date Tags for Anonymization}
\format{A \code{data.frame} with 2 columns
\describe{
\item{tag}{the tag identifier for the field }
\item{name}{the name of the field }
}}
\usage{
date_tags
}
\description{
A \code{data.frame} of tags of dates, useful for DICOM
anonymization.
}
\keyword{datasets}
