% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcm2pdf.R
\name{dcm2pdf}
\alias{dcm2pdf}
\title{DICOM to PDF}
\usage{
dcm2pdf(file, opts = "", outfile = NULL)
}
\arguments{
\item{file}{DICOM file name}

\item{opts}{options to pass to \code{dcm2pdf}}

\item{outfile}{output PDF filename}
}
\value{
Character vector of output filename
}
\description{
Wrapper of \code{dcm2pdf} from DCMTK
}

