% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcm2xml.R
\name{dcm2xml}
\alias{dcm2xml}
\title{DICOM to XML}
\usage{
dcm2xml(file, opts = "", outfile = NULL)
}
\arguments{
\item{file}{DICOM file name}

\item{opts}{options to pass to \code{dcm2xml}}

\item{outfile}{output XML filename}
}
\value{
Character vector of output filename
}
\description{
Wrapper of \code{dcm2xml} from DCMTK
}

