% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcm_anon.R
\name{dcm_anon}
\alias{dcm_anon}
\title{DICOM Anonymization}
\usage{
dcm_anon(
  file,
  rem_tags = dcmtk::removal_tags$tag,
  mod_tags = NULL,
  insert_tags = NULL,
  new_uids = TRUE,
  new_date = NULL,
  new_time = NULL,
  verbose_flag = FALSE,
  ...
)
}
\arguments{
\item{file}{DICOM Image}

\item{rem_tags}{tags to remove}

\item{mod_tags}{tags to modify.  Must be a character vector, with the
names of that vector as the tags.}

\item{insert_tags}{tags to insert.  Must be a character vector, with the
names of that vector as the tags.}

\item{new_uids}{Should new unique identifiers (IDs) be generated}

\item{new_date}{date to be change date fields to.}

\item{new_time}{time to be change time fields to.}

\item{verbose_flag}{print diagnostic messages, using
\code{--verbose}}

\item{...}{Additional arguments to add to \code{\link{dcmodify}}}
}
\value{
Result of \code{dcmodify} command
}
\description{
Anonymizes DICOM images
}
\examples{
\dontrun{
file = "~/Desktop/000000.dcm"
anon = dcm_anon(file, new_date = "20060101")
new_hd = read_dicom_header(anon)
}
}
