% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcm_subtract_date.R
\name{dcm_subtract_date}
\alias{dcm_subtract_date}
\title{DICOM Anonymization}
\usage{
dcm_subtract_date(file, base_date, origin = as.Date("19000101", format =
  "\%Y\%m\%d"), date_tags = dcmtk::date_tags$tag)
}
\arguments{
\item{file}{DICOM Image}

\item{base_date}{baseline date to subtract dates from.
The new date is
\code{date - base_date + origin}.}

\item{origin}{origin to add back in.  The new date is
\code{date - base_date + origin}.}

\item{date_tags}{tags to modify}
}
\value{
Output filename
}
\description{
Anonymizes DICOM images
}
\note{
If no tags are changed, then a warning will be printed.
}
\examples{
\dontrun{
file = "~/Desktop/000000.dcm"
base_date = lubridate::ymd("2004-06-10")
origin = as.Date("19000101", format = "\%Y\%m\%d")
tfile = tempfile()
file.copy(file, tfile)
anon = dcm_subtract_date(tfile, base_date = base_date)
hdr = read_dicom_header(anon)
}
}
