% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcmdjpeg.R
\name{dcmdjpeg}
\alias{dcmdjpeg}
\title{Decode JPEG-compressed DICOM file}
\usage{
dcmdjpeg(file, outfile = NULL, ...)
}
\arguments{
\item{file}{DICOM file name}

\item{outfile}{output DICOM file name}

\item{...}{options passed to \code{\link{dcmtk_cmd}}}
}
\value{
Character output file name
}
\description{
Wrapper of \code{dcmdjpeg} from DCMTK
}
