% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcmdrle.R
\name{dcmdrle}
\alias{dcmdrle}
\title{Decode DICOM with RLE}
\usage{
dcmdrle(file, opts = "", outfile = NULL)
}
\arguments{
\item{file}{DICOM file name}

\item{opts}{options to pass to \code{dcmdrle}}

\item{outfile}{output DICOM filename}
}
\value{
Character vector of output filename
}
\description{
Wrapper of \code{dcmdrle} from DCMTK to
decode DICOM with run-length encoding
}

