% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcmj2pnm.R
\name{dcmj2pnm}
\alias{dcmj2pnm}
\title{Convert DICOM images to PGM/PPM, PNG, JPEG or BMP using
\code{dcmdump} from DCMTK}
\usage{
dcmj2pnm(file, outfile = tempfile(fileext = ".png"),
  opts = "--write-png", ...)
}
\arguments{
\item{file}{DICOM file name}

\item{outfile}{output file name passed to \code{\link{dcmtk_cmd}}}

\item{opts}{opts passed to \code{\link{dcmtk_cmd}}}

\item{...}{options passed to \code{\link{dcmtk_cmd}}
\code{intern}}
}
\value{
Character vector of information
}
\description{
Convert DICOM images to PGM/PPM, PNG, JPEG or BMP using
\code{dcmdump} from DCMTK
}
\examples{
file = system.file("extdata", "example.dcm", package = "dcmtk")
png_file = dcmj2pnm(file)
img = png::readPNG(png_file)
plot(1:2, type='n')
image(img)

}
