% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcmtk_cmd.R
\name{dcmtk_cmd}
\alias{dcmtk_cmd}
\title{Get path for DCMTK command}
\usage{
dcmtk_cmd(cmd)
}
\arguments{
\item{cmd}{DCMTK command name}
}
\value{
Logical \code{TRUE} if command is available.
}
\description{
Uses \code{Sys.which} to check if command is in the PATH.
}
\examples{
dcmtk_cmd("dcmodify")
}

