% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcmtk_cmd_path.R
\name{dcmtk_cmd_path}
\alias{dcmtk_cmd_path}
\title{Get path for DCMTK command}
\usage{
dcmtk_cmd_path(cmd)
}
\arguments{
\item{cmd}{DCMTK command name}
}
\value{
Logical \code{TRUE} if command is available.
}
\description{
Uses \code{Sys.which} to check if command is in the PATH.
}
\examples{
if (!have_dcmtk_cmd("dcmodify")) {
install_dir = tempdir()
options(dcmtk.path = install_dir)
  res = try({
    install_dcmtk(install_dir = install_dir)
  })
  if (inherits(res, "try-error")) {
    res = FALSE
  }
  if (!res) {
    source_install_dcmtk(install_dir = install_dir)
  }
}
dcmtk_cmd("dcmodify")
}
