% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcmtk_help.R
\name{dcmtk_help}
\alias{dcmtk_help}
\title{Wrapper for getting DCMTK help}
\usage{
dcmtk_help(cmd, help.arg = "--help")
}
\arguments{
\item{cmd}{DCMTK command}

\item{help.arg}{Argument to print help, usually "--help"}
}
\value{
Prints help output and returns output as character vector
}
\description{
This function takes in the function and returns the
help from DCMTK for that function
}

