% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicom_tags.R
\docType{data}
\name{dicom_tags}
\alias{dicom_tags}
\title{DICOM Tags}
\format{
A \code{data.frame} with 6 columns
\describe{
\item{tag}{the tag identifier for the field }
\item{name}{the name of the field }
\item{keyword}{ no space version of name - a better lookup ID}
\item{vr}{value representation}
\item{vm}{value multiplier}
\item{retired}{indication of retired tag}
}
}
\usage{
dicom_tags
}
\description{
A \code{data.frame} of tags to look up in DICOM
header.  Not an exhaustive list, but extensive.
}
\note{
Adapted from
\url{http://dicom.nema.org/dicom/2013/output/chtml/part06/chapter_6.html}
}
\keyword{datasets}
