% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/have_dcmtk_cmd.R
\name{have_dcmtk_cmd}
\alias{check_dcmtk_cmd}
\alias{have_dcmtk_cmd}
\title{Logical check if a DCMTK function is available}
\usage{
have_dcmtk_cmd(cmd)

check_dcmtk_cmd(cmd)
}
\arguments{
\item{cmd}{DCMTK command name}
}
\value{
Logical \code{TRUE} if command is available.
}
\description{
Uses \code{Sys.which} to check if command is in the PATH.
}
\examples{
have_dcmtk_cmd("dcmodify")
}

