% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_dcmtk.R
\name{install_dcmtk}
\alias{install_dcmtk}
\title{Install DCMTK}
\usage{
install_dcmtk(type = c("osx", "linux_static", "linux_dynamic",
  "windows"), force = FALSE, version = c("3.6.3", "3.6.2", "3.6.0"),
  arch = c("x86_64", "i686"), install_dir = system.file(package =
  "dcmtk"))
}
\arguments{
\item{type}{Which format of dcmtk should be downloaded.  If not specified,
it will be chosen from the platform}

\item{force}{Should installing be forced even if folders exist?}

\item{version}{version of DCMTK to install}

\item{arch}{architecture to install, either 64 or 32}

\item{install_dir}{Installation Directory}
}
\value{
Logical
}
\description{
Installs the DICOM ToolKit
}
\examples{
install_dir =  tempfile()
dir.create(install_dir, showWarnings = FALSE, recursive = TRUE)
install_dcmtk(install_dir = install_dir)
}
