% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_hdr.R
\name{parse_hdr}
\alias{parse_hdr}
\title{Parse DICOM header}
\usage{
parse_hdr(hdr, convert_non_ascii = TRUE)
}
\arguments{
\item{hdr}{Character vector from \code{\link{dcmdump}}}

\item{convert_non_ascii}{Should [iconv] be run before parsing?  This
can help failures in UTF-8 issues, but may alter the output from
[dcmdump()]}
}
\value{
\code{data.frame} of tags and values
}
\description{
Parses a DICOM header to a \code{data.frame}
}
\examples{
file = system.file("extdata", "example.dcm", package = "dcmtk")
hdr = dcmdump(file)
parsed = parse_hdr(hdr)
}
