% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removal_tags.R
\docType{data}
\name{removal_tags}
\alias{removal_tags}
\title{Removal Tags for Anonymization}
\format{A \code{data.frame} with 2 columns
\describe{
\item{tag}{the tag identifier for the field }
\item{name}{the name of the field }
}}
\usage{
removal_tags
}
\description{
A \code{data.frame} of tags to remove when doing DICOM
anonymization.
}
\keyword{datasets}
