% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide_hdr.R
\name{wide_hdr}
\alias{wide_hdr}
\title{Reshape the Header files}
\usage{
wide_hdr(hdr, keep_tags = NULL)
}
\arguments{
\item{hdr}{An output from \code{\link{read_dicom_header}}}

\item{keep_tags}{tags to keep in the data}
}
\value{
A wide \code{data.frame} of the information needed
for filtering
}
\description{
Reshape the Header files
}
