% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcmdump.R
\name{dcmdump}
\alias{dcmdump}
\title{DICOM Dump}
\usage{
dcmdump(file, ...)
}
\arguments{
\item{file}{DICOM file name}

\item{...}{options passed to \code{\link{dcmtk_cmd}}, other than
\code{intern}}
}
\value{
Character vector of information
}
\description{
Wrapper of \code{dcmdump} from DCMTK
}
