% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcmj2pnm.R
\name{dcmj2pnm}
\alias{dcmj2pnm}
\title{Convert DICOM images to PGM/PPM, PNG, JPEG or BMP using
\code{dcmdump} from DCMTK}
\usage{
dcmj2pnm(file, outfile = tempfile(fileext = ".png"), opts = "--write-png", ...)
}
\arguments{
\item{file}{DICOM file name}

\item{outfile}{output file name passed to \code{\link{dcmtk_cmd}}}

\item{opts}{opts passed to \code{\link{dcmtk_cmd}}}

\item{...}{options passed to \code{\link{dcmtk_cmd}}
\code{intern}}
}
\value{
Character vector of information
}
\description{
Convert DICOM images to PGM/PPM, PNG, JPEG or BMP using
\code{dcmdump} from DCMTK
}
\examples{
if (!have_dcmtk_cmd("dcmj2pnm")) {
  install_dir = tempdir()
  res = try({
    install_dcmtk(install_dir = install_dir)
  })
  if (inherits(res, "try-error")) {
    res = FALSE
  }
  if (!res) {
    source_install_dcmtk(install_dir = install_dir)
  }
  options(dcmtk.path = install_dir)
}
file = system.file("extdata", "example.dcm", package = "dcmtk")
png_file = dcmj2pnm(file)
print(png_file)
print(file.exists(png_file))
print(normalizePath(png_file))
img = png::readPNG(png_file)
plot(1:2, type='n')
image(img)

}
