% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_dcmtk.R
\name{install_dcmtk}
\alias{install_dcmtk}
\title{Install DCMTK}
\usage{
install_dcmtk(type = c("osx", "linux_static", "linux_dynamic",
  "windows"), force = FALSE)
}
\arguments{
\item{type}{Which format of dcmtk should be downloaded.  If not specified,
it will be chosen from the platform}

\item{force}{Should installing be forced even if folders exist?}
}
\value{
Logical
}
\description{
Installs the DICOM ToolKit
}
