% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_hdr.R
\name{parse_hdr}
\alias{parse_hdr}
\title{Parse DICOM header}
\usage{
parse_hdr(hdr)
}
\arguments{
\item{hdr}{Character vector from \code{\link{dcmdump}}}
}
\value{
\code{data.frame} of tags and values
}
\description{
Parses a DICOM header to a \code{data.frame}
}
