% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dicom_header.R
\name{read_dicom_header}
\alias{dcmhd}
\alias{read_dicom_header}
\title{Read DICOM File}
\usage{
read_dicom_header(file)

dcmhd(file)
}
\arguments{
\item{file}{Name of DICOM file}
}
\value{
Character vector of header information
}
\description{
Reads in the DICOM header from a file
}

