% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dicom_header.R
\name{read_dicom_header}
\alias{dcmhd}
\alias{read_dicom_header}
\title{Read DICOM File}
\usage{
read_dicom_header(file, replace_names = FALSE, add_opts = "",
  recursive = FALSE)

dcmhd(file)
}
\arguments{
\item{file}{Name of DICOM file}

\item{replace_names}{logical indicating if unknown tag names should be
inferred from \code{dicom_tags}}

\item{add_opts}{additional options to pass to \code{\link{dcmdump}}.
The flags already added are
\code{-q --print-all --load-short --print-filename}}

\item{recursive}{logical indicating if the \code{--recurse} flag be passed to
\code{\link{dcmdump}}}
}
\value{
Character vector of header information
}
\description{
Reads in the DICOM header from a file
}

