% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dicom_header.R
\name{read_dicom_header}
\alias{read_dicom_header}
\alias{dcmhd}
\title{Read DICOM File}
\usage{
read_dicom_header(file = "*.dcm", replace_names = FALSE, add_opts = "",
  recursive = FALSE, path = ".")

dcmhd(file)
}
\arguments{
\item{file}{DICOM input file or directory.  If \code{recursivee = TRUE},
then this will be the pattern to match within \code{path}}

\item{replace_names}{logical indicating if unknown tag names should be
inferred from \code{dicom_tags}}

\item{add_opts}{additional options to pass to \code{\link{dcmdump}}.
The flags already added are
\code{-q --print-all --load-short --print-filename}}

\item{recursive}{logical indicating if the \code{--recurse} flag be passed to
\code{\link{dcmdump}}}

\item{path}{if \code{recursive = TRUE}, then this will the path scanned.}
}
\value{
Character vector of header information
}
\description{
Reads in the DICOM header from a file
}
